Add ntfy request privacy options
Ntfy accepts a couple of privacy parameters that allow one to disable
server caching, and broadcasting with Firebase. These have been added
as CLI flags to allow users to request opting-out of these features.
I may make them opt-in in the future, but as they're generally
beneficial I've chosen to leave them as is for now.