Modify notifier API, config opt-in
Instead of accepting the configuration optionally in `new`, we now
accept configuration later in `configure`. This makes it such that we
don't have to awkwardly pass in `None` in `new` if we don't want to add
configuration parameters on initialization. This, admittedly is unlikely
to see much usage in this project, but it aligns more with design best
practices.